+++ packages/firebase/src/lib/common/model/permission/permission.service.grant.ts
@@ -1,3 +1,4 @@
+import { FirestoreDocument } from './../../firestore/accessor/document';
 import { fullAccessRoleMap, GrantedRoleMap, noAccessRoleMap } from '@dereekb/model';
 import { AsyncDecisionFunction, Getter, GetterOrValue, getValueFromGetter, Maybe, PromiseOrValue } from '@dereekb/util';
 import { FirebaseModelContext } from '../context';
@@ -7,7 +8,7 @@ import { UserRelated } from '../../../model/user';
 /**
  * Convenience function that checks the input context if the user is an admin or grants all roles.
  */
-export const grantFullAccessIfAdmin: GeneralGrantRolesIfFunction = grantModelRolesIfAdminFunction(fullAccessRoleMap);
+export const grantFullAccessIfAdmin: GeneralGrantRolesIfFunction<FirebaseModelContext> = grantModelRolesIfAdminFunction(fullAccessRoleMap);
 
 export function grantModelRolesIfAdmin<R extends string = string>(context: FirebaseModelContext, rolesToGrantToAdmin: GetterOrValue<GrantedRoleMap<R>>, otherwise?: GrantRolesOtherwiseFunction<R>): PromiseOrValue<GrantedRoleMap<R>> {
   return grantModelRolesIfAdminFunction(rolesToGrantToAdmin)(context, otherwise);
@@ -34,15 +35,22 @@ export function grantModelRolesIfAdminFunction<R extends string = string>(rolesT
 export const isAdminInFirebaseModelContext: AsyncDecisionFunction<FirebaseModelContext> = (context: FirebaseModelContext) => context.auth?.isAdmin() ?? false;
 
 // MARK: User Related
-export type UserRelatedModelFirebaseModelContext<T extends UserRelated = UserRelated> = {
+export type UserRelatedModelFirebaseModelContext<T extends UserRelated = UserRelated> = UserRelatedModelFirebaseModelContextModelInput<T> | UserRelatedModelFirebaseModelContextDocumentInput<T>;
+
+export type UserRelatedModelFirebaseModelContextModelInput<T extends UserRelated = UserRelated> = {
   model: T;
+  context: FirebaseModelContext;
+};
+
+export type UserRelatedModelFirebaseModelContextDocumentInput<T extends UserRelated = UserRelated> = {
+  document: FirestoreDocument<T>;
   context: FirebaseModelContext;
 };
 
 /**
  * Convenience function that checks the input context if the user is related to the model by uid.
  */
-export const grantFullAccessIfAuthUserRelated: GeneralGrantRolesIfFunction = grantModelRolesIfAdminFunction(fullAccessRoleMap);
+export const grantFullAccessIfAuthUserRelated: GeneralGrantRolesIfFunction<UserRelatedModelFirebaseModelContext<UserRelated>> = grantModelRolesIfAuthUserRelatedModelFunction(fullAccessRoleMap);
 
 /**
  * Creates a GrantRolesIfFunction that grants roles if the user is related to the model by uid.
@@ -62,14 +70,35 @@ export function grantModelRolesIfAuthUserRelatedModelFunction<T extends UserRela
  * @param context
  * @returns
  */
-export const isOwnerOfUserRelatedModelInFirebaseModelContext: AsyncDecisionFunction<UserRelatedModelFirebaseModelContext<UserRelated>> = ({ context, model }: UserRelatedModelFirebaseModelContext) => context.auth?.uid === model.uid;
+export const isOwnerOfUserRelatedModelInFirebaseModelContext: AsyncDecisionFunction<UserRelatedModelFirebaseModelContext<UserRelated>> = async (context: UserRelatedModelFirebaseModelContext) => {
+  let decision = false;
+  const auth = context.context.auth;
+
+  if (auth != null) {
+    let model: Maybe<UserRelated> = (context as UserRelatedModelFirebaseModelContextModelInput).model;
+
+    if (!model) {
+      const document = (context as UserRelatedModelFirebaseModelContextDocumentInput).document;
+
+      if (document) {
+        model = (await document.accessor.get()).data();
+      }
+    }
+
+    if (model) {
+      decision = auth.uid === model.uid;
+    }
+  }
+
+  return decision;
+};
 
 // MARK: Grant Roles
 /**
  * Grants the configured roles if the decision is made about the context. Otherwise, returns a NoAccessRoleMap.
  */
 export type GrantRolesOnlyIfFunction<C, R extends string = string> = (context: C) => Promise<GrantedRoleMap<R>>;
-export type GeneralGrantRolesOnlyIfFunction = <C, R extends string = string>(context: C) => Promise<GrantedRoleMap<R>>;
+export type GeneralGrantRolesOnlyIfFunction<C> = <R extends string = string>(context: C) => Promise<GrantedRoleMap<R>>;
 
 /**
  * Creates a GrantRolesOnlyIfFunction
@@ -87,7 +116,7 @@ export function grantModelRolesOnlyIfFunction<C, R extends string = string>(gran
  * Grants the configured roles if the decision is made about the context. Otherwise, invokes the otherwise function if available, or returns a NoAccessRoleMap.
  */
 export type GrantRolesIfFunction<C, R extends string = string> = (context: C, otherwise?: GrantRolesOtherwiseFunction<R>) => Promise<GrantedRoleMap<R>>;
-export type GeneralGrantRolesIfFunction = <C, R extends string = string>(context: C, otherwise?: GrantRolesOtherwiseFunction<R>) => Promise<GrantedRoleMap<R>>;
+export type GeneralGrantRolesIfFunction<C> = <R extends string = string>(context: C, otherwise?: GrantRolesOtherwiseFunction<R>) => Promise<GrantedRoleMap<R>>;
 
 /**
  * Used as the "else" statement for grantModelRolesIfFunction.