+++ packages/util/src/lib/promise/use.ts
@@ -1,20 +1,18 @@
 import { asGetter, GetterOrValue } from '../getter';
-import { PromiseOrValue } from './promise';
-
-export type UsePromiseUseFunction<I, O> = (value: I) => PromiseOrValue<O>;
+import { UseAsync } from '../value';
 
 /**
- * Uses a value returned from a promise.
+ * Uses a cached promise value.
  */
-export type UsePromiseFunction<I> = <O>(useFn: UsePromiseUseFunction<I, O>) => Promise<O>;
+export type UsePromiseFunction<I> = <O>(useFn: UseAsync<I, O>) => Promise<O>;
 
 /**
- * Creates a UsePromiseFactory.
+ * Creates a UsePromiseFunction.
  *
  * @param input
  * @returns
  */
 export function usePromise<I>(input: GetterOrValue<Promise<I>>): UsePromiseFunction<I> {
   const _getter = asGetter(input);
-  return <O>(useFn: UsePromiseUseFunction<I, O>) => _getter().then(useFn) as Promise<O>;
+  return <O>(useFn: UseAsync<I, O>) => _getter().then(useFn) as Promise<O>;
 }